refactor: move memory module
