+++ agent/__init__.py
@@ -26,7 +26,7 @@ from .tools import (
 from .utils.helpers import limit_chars
 from .utils.speech import transcribe_audio
 from .vm import LinuxVM
-from .memory import get_memory, set_memory, edit_memory
+from .utils.memory import get_memory, set_memory, edit_memory
 
 __all__ = [
     "ChatSession",
