+++ agent/utils/__init__.py
@@ -1,8 +1,12 @@
 """Utility functions exposed at package level."""
 
 from .speech import transcribe_audio
+from .memory import get_memory, set_memory, edit_memory
 
 __all__ = [
-    "transcribe_audio"
+    "transcribe_audio",
+    "get_memory",
+    "set_memory",
+    "edit_memory",
 ]
