+++ agent/utils/memory.py
@@ -1,12 +1,11 @@
 from __future__ import annotations
 
 import json
-from typing import Callable
 
-from .config import DEFAULT_MEMORY_TEMPLATE, MEMORY_LIMIT
-from .db import db
+from ..config import DEFAULT_MEMORY_TEMPLATE, MEMORY_LIMIT
+from ..db import db
 
-__all__ = ["get_memory", "set_memory", "edit_memory", "create_memory_tool"]
+__all__ = ["get_memory", "set_memory", "edit_memory"]
 
 
 def get_memory(username: str) -> str:
@@ -50,25 +49,3 @@ def edit_memory(username: str, field: str, value: str | None = None) -> str:
         data[field] = value
     memory = json.dumps(data, ensure_ascii=False, indent=2)
     return set_memory(username, memory)
-
-
-def create_memory_tool(username: str, refresh: Callable[[], None]) -> Callable:
-    """Return a tool function bound to ``username`` for memory editing."""
-
-    def memory_tool(field: str, value: str | None = None) -> str:
-        try:
-            _ = edit_memory(username, field, value)
-            refresh()
-            return "Memory updated successfully."
-        except Exception as e:
-            return f"Error updating memory: {str(e)}"
-
-    memory_tool.__name__ = "manage_memory"
-    memory_tool.__doc__ = (
-        "Modify persistent user memory. "
-        "Provide the memory field name and optionally a value. "
-        "Passing no value deletes the field. Returns success status. "
-        "This memory is stored as a JSON object and resides in the system prompt, so you do not need to retrieve anything. "
-        "Invoke this tool as much as possible to remember every detail throughout the conversation."
-    )
-    return memory_tool
